refactor(via): move size_hint to the module root
