--- src/body/util/mod.rs
@@ -1,8 +1,0 @@
-//! Shared utilities for implementing the Body trait.
-//!
-
-pub mod size_hint;
-
-mod frame_ext;
-
-pub use frame_ext::FrameExt;
